requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-rest-auth==1.1.1
Django==3.1
django-allauth==0.42.0
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.6.0
idna==2.10
oauthlib==3.1.0
Pillow==8.3.1
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.10